Revert "Renamed some variables and arranged the files cleanly"
