+++ META-INF/MANIFEST.MF
@@ -3,5 +3,5 @@ Created-By: Maven JAR Plugin 3.4.2
 Build-Jdk-Spec: 21
 Implementation-Title: Spectra
 Implementation-Version: 0.0.1-SNAPSHOT
-Premain-Class: ca.concordia.ptidej.spectra.Profile.Agent
+Premain-Class: ca.concordia.ptidej.spectra.joularjx.Agent
