+++ pom.xml
@@ -128,7 +128,7 @@
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                                     <manifestEntries>
-                                        <Premain-Class>ca.concordia.ptidej.spectra.Profile.Agent</Premain-Class>
+                                        <Premain-Class>ca.concordia.ptidej.spectra.joularjx.Agent</Premain-Class>
                                     </manifestEntries>
 
                                 </transformer>
